requirements.txt
streamlit
numpy
pandas
PyPDF2
langchain
openai
python-dotenv
toml
google-cloud-storage>=1.0.0
google-cloud-firestore
google-auth
